wedding_project/wedding_project/requirements.txt
asgiref==3.2.10
azure-common==1.1.25
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.2
django-storages==1.10.1
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.0.4
idna==2.10
Pillow==7.2.0
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.0
static3==0.7.0
urllib3==1.25.10
whitenoise==5.2.0
